Text ui component working in guest
