+++ crates/guest_bridge/src/lib.rs
@@ -3,6 +3,9 @@ pub use ambient_ecs as ecs;
 
 #[cfg(feature = "native")]
 pub mod components {
+    pub mod app {
+        pub use ambient_core::{name, ui_scene};
+    }
     pub mod ecs {
         pub use ambient_core::hierarchy::{children, parent};
     }
@@ -10,7 +13,11 @@ pub mod components {
         pub use ambient_core::transform::{local_to_parent, local_to_world, mesh_to_local, mesh_to_world, rotation, scale, translation};
     }
     pub mod ui {
-        pub use ambient_layout::*;
+        pub use ambient_layout::{height, min_height, min_width, width};
+        pub use ambient_text::{font_size, text};
+    }
+    pub mod rendering {
+        pub use ambient_renderer::color;
     }
 }
