+++ crates/ui_components/Cargo.toml
@@ -6,6 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ambient_element = { path = "../element" }
-ambient_guest_bridge = { path = "../guest_bridge" }
+ambient_element = { path = "../element", default-features = false }
+ambient_guest_bridge = { path = "../guest_bridge", default-features = false }
 glam = { workspace = true }
