+++ crates/ui_components/src/lib.rs
@@ -17,3 +17,12 @@ pub fn UIBase(_: &mut Hooks) -> Element {
         .init(width(), 0.)
         .init(height(), 0.)
 }
+
+/// This only exists so that we can implement From<String> for Text, and then use it in
+/// for instance Button
+pub struct UIElement(pub Element);
+impl From<Element> for UIElement {
+    fn from(el: Element) -> Self {
+        Self(el)
+    }
+}
